examples/projects/websocket/pyproject.toml
[tool.poetry]
name = "unrealon-websocket"
version = "1.0.0"
description = "Unrealon Cloud IDE - WebSocket RPC Server"
authors = ["Unrealon Team <dev@unrealon.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
# django-ipc from PyPI
django-ipc = "*"
# Additional dependencies
python-dotenv = "^1.0.0"
uvloop = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^0.25.0"
black = "^25.9.0"
mypy = "^1.18.0"
fakeredis = "^2.31.3"


[tool.poetry.group.local.dependencies]
django-ipc = {path = "/Users/markinmatrix/djangoipc", develop = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
strict = true
